fix: package circle reference
