+++ packages/types/package.json
@@ -5,7 +5,6 @@
   "type": "module",
   "main": "./src/index.ts",
   "dependencies": {
-    "@lobechat/model-runtime": "workspace:*",
     "@lobechat/python-interpreter": "workspace:*",
     "@lobechat/web-crawler": "workspace:*",
     "@lobehub/chat-plugin-sdk": "^1.32.4",
