+++ packages/types/src/agentRuntime.ts
@@ -1,0 +1,47 @@
+/* eslint-disable sort-keys-fix/sort-keys-fix */
+// ******* Runtime Biz Error ******* //
+export const AgentRuntimeErrorType = {
+  AgentRuntimeError: 'AgentRuntimeError', // Agent Runtime 模块运行时错误
+  LocationNotSupportError: 'LocationNotSupportError',
+
+  QuotaLimitReached: 'QuotaLimitReached',
+  InsufficientQuota: 'InsufficientQuota',
+
+  ModelNotFound: 'ModelNotFound',
+
+  PermissionDenied: 'PermissionDenied',
+  ExceededContextWindow: 'ExceededContextWindow',
+
+  InvalidProviderAPIKey: 'InvalidProviderAPIKey',
+  ProviderBizError: 'ProviderBizError',
+
+  InvalidOllamaArgs: 'InvalidOllamaArgs',
+  OllamaBizError: 'OllamaBizError',
+  OllamaServiceUnavailable: 'OllamaServiceUnavailable',
+
+  InvalidBedrockCredentials: 'InvalidBedrockCredentials',
+  InvalidVertexCredentials: 'InvalidVertexCredentials',
+  StreamChunkError: 'StreamChunkError',
+
+  InvalidGithubToken: 'InvalidGithubToken',
+
+  ConnectionCheckFailed: 'ConnectionCheckFailed',
+
+  // ******* Image Generation Error ******* //
+  ProviderNoImageGenerated: 'ProviderNoImageGenerated',
+
+  InvalidComfyUIArgs: 'InvalidComfyUIArgs',
+  ComfyUIBizError: 'ComfyUIBizError',
+  ComfyUIServiceUnavailable: 'ComfyUIServiceUnavailable',
+  ComfyUIEmptyResult: 'ComfyUIEmptyResult',
+  ComfyUIUploadFailed: 'ComfyUIUploadFailed',
+  ComfyUIWorkflowError: 'ComfyUIWorkflowError',
+  ComfyUIModelError: 'ComfyUIModelError',
+
+  /**
+   * @deprecated
+   */
+  NoOpenAIAPIKey: 'NoOpenAIAPIKey',
+} as const;
+export type ILobeAgentRuntimeErrorType =
+  (typeof AgentRuntimeErrorType)[keyof typeof AgentRuntimeErrorType];
