+++ packages/types/src/index.ts
@@ -36,6 +36,7 @@ export * from './user';
 export * from './userMemory';
 // FIXME: I think we need a refactor for the "openai" types
 // it more likes the UI message payload
+export * from './agentRuntime';
 export * from './openai/chat';
 export * from './openai/plugin';
 export * from './subscription';
